remove InMemoryEventStoreQueryTrait
